requirements.txt
asgiref==3.4.1
Django==3.2.7
pytz==2021.1
sqlparse==0.4.1

djangorestframework===3.12.4
django-debug-toolbar==3.2.2
coverage==5.5
factory-boy==3.2.1

drf-yasg==1.20.0